Finish database testing framework
database.UnitTests is an array of functions and names of units tests
that should be run. The postgres tests then import them, and in a loop
initialise the database, run the test as a sub-test, then delete every
table from the database (so that the db is empty for the next test).