Use healthcheck to guarentee postgres ready
If we only set depends_on to postgres, then the control container will
be run as soon as the postgres one is running, but if the postgres
container has some initialisation work to do, this may lead to the
previous issue reemerging.

Instead, use `pg_isready` to ensure that that the service is healthy
before starting the control container.

I've set an interval of 10s, down from the default of 30s. This has two
effects. The first healthcheck isn't run until after interval seconds
have passed (putting a lower bound on the start-up time), and will
continue to be run every interval settings (so if it's).