Prevent satellite from dying instantly
Go main threads don't wait for goroutines to finish, so we need to
loop somewhere in the main thread, and the collect/publish stats loop
seems as good a place as anywhere.

Also make the satellite collect and publish stats instantly instead of
waiting for the first tick. This means we get to see that the satellite
is working much faster for larger intervals.