Update implementation of Postgres database
With the new datatype, we now need to pull out machines and put them
into groups, rather than return them as is.

We have a default group we put machines that either have a null or empty
group into.

I don't believe we can combine the two for loops into one in
`LatestData()`, because we can't start a new query before the rows of
the last query have been closed (happens automatically once .Next()
returns false)