Update in-memory implementation
Inmem was not previously tracking machine info (groups, notes, etc.), so
add a new map to track that.

I've tried to keep other changes to the inmemory struct to a minimum, so
latest data involves a fair bit of zipping other data structures
together. I've tried to reduce the pain of later changes with reflect,
but more could probably be done.