Update tests for new data types
Biggest change here is making it so statsNear takes both the
GPUStatSample and GPUInfo we put in to compare to the one combined
struct. A bit of reflect here could def clean things up, but vim macros
were easier and I wanted to get it done.

Other than that, it's mainly changing data types and the way traversing
happens

More tests could definately be written