Fix date we test downsampling
The precise number of downsamples we get seems to vary depending on when
the test is run (it's started failing since the start of the month.

Therefore, as a temporary patch, fix the rundate of the downsample test
to be the beginning of Febuary, this should cover us for now.