Ask for login when we try to change to admin panel
Rather than click on login, type in username/password, then click on
admin panel, when we click on admin panel, if we're not already logged
in, then pop up the panel.

Had to make some changes to login so we know whether it succeeded or not
so we know whether to switch to the admin panel.

Had to add some state to the navbar, the current index, so we can
control it programmatically.